+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.7.0
 
 - Mobile, Web: Added priority selection buttons to task form.
+- Mobile, Web: Added support for contexts.
 - Mobile, Web: Fixed project search bug at task form.
 - Mobile, Web: Fixed bug where task IDs become wrong.
 - Web: Created plaintext page.